修复db参数类型错误

+++ smartcontract/wasmservice/database.go
@@ -5,13 +5,13 @@ import(
 )
 
 //for c api: int db_put(char* key, uint32 k_len, char *value, uint32 v_len)
-func(ws *WasmService)db_put(proc *exec.Process, key, k_len, value, v_len int)int32{
+func(ws *WasmService)db_put(proc *exec.Process, key, k_len, value, v_len int32)int32{
 	var k_msg, v_msg []byte
-	err := proc.ReadAt(k_msg, key, k_len)
+	err := proc.ReadAt(k_msg, int(key), int(k_len))
 	if err != nil{
 		return -1
 	}
-	err = proc.ReadAt(v_msg, value, v_len)
+	err = proc.ReadAt(v_msg, int(value), int(v_len))
 	if err != nil{
 		return -1
 	}
@@ -22,10 +22,10 @@ func(ws *WasmService)db_put(proc *exec.Process, key, k_len, value, v_len int)int
 	return 0
 }
 
-//for c api:char *db_get(char* key, uint32 k_len, char *value, uint32 v_len)
-func(ws *WasmService)db_get(proc *exec.Process, key, k_len, value, v_len int)int32{
+//for c api: int db_get(char* key, uint32 k_len, char *value, uint32 v_len)
+func(ws *WasmService)db_get(proc *exec.Process, key, k_len, value, v_len int32)int32{
 	var k_msg []byte
-	err := proc.ReadAt(k_msg, key, k_len)
+	err := proc.ReadAt(k_msg, int(key), int(k_len))
 	if err != nil{
 		return -1
 	}
@@ -33,10 +33,10 @@ func(ws *WasmService)db_get(proc *exec.Process, key, k_len, value, v_len int)int
 	if err != nil{
 		return -1
 	}
-	if v_len > len(v_msg){
-		v_len = len(v_msg)
+	if int(v_len) > len(v_msg){
+		   v_len = int32(len(v_msg))
 	}
-	err = proc.WriteAt(v_msg[:], value, v_len)
+	err = proc.WriteAt(v_msg[:], int(value), int(v_len))
 	if err != nil{
 		return -1
 	}
